fix: cte within cte
